Fix failing unit test.
