+++ app/actions/handlers.py
@@ -1,11 +1,14 @@
 # actions/handlers.py
 import logging
 from datetime import datetime, timedelta
+from typing import Tuple
 
 from app.actions.helpers import Environment, get_er_token_and_site
 from app.services.action_scheduler import crontab_schedule
 from app.services.activity_logger import activity_logger, log_activity
 from app.services.gundi import send_observations_to_gundi
+from app.services.utils import find_config_for_action
+from gundi_core import schemas
 from gundi_core.events import LogLevel
 from gundi_core.schemas.v2 import Integration
 from gundi_client_v2 import GundiClient
@@ -17,6 +20,10 @@ logger = logging.getLogger(__name__)
 
 
 LOAD_BATCH_SIZE = 100
+
+headers = {
+    "Authorization": f"Bearer ",
+}
 
 
 async def action_auth(integration: Integration, action_config: AuthenticateConfig):
@@ -187,3 +194,34 @@ async def action_pull_observations_24_hour_sync(
 def generate_batches(iterable, n=LOAD_BATCH_SIZE):
     for i in range(0, len(iterable), n):
         yield iterable[i : i + n]
+
+
+async def get_er_token_and_site(
+    integration: Integration, environment: Environment
+) -> Tuple[str, str]:
+    """
+    Get the ER token and site for the given integration and environment
+
+    :param integration: Integration object
+    :param environment: Environment enum
+    :return: Tuple of ER token and site
+    """
+    _client = GundiClient()
+    connection_details = await _client.get_connection_details(integration.id)
+
+    destination = (
+        destination
+        for destination in connection_details.destinations
+        if environment.value in destination.name
+    ).__next__()
+
+    destination_details = await _client.get_integration_details(destination.id)
+    auth_config = find_config_for_action(
+        configurations=destination_details.configurations,
+        action_id="auth",
+    )
+
+    auth_config = schemas.v2.ERAuthActionConfig.parse_obj(auth_config.data)
+    if auth_config:
+        return auth_config.token, destination_details.base_url
+    return None, None
